server/requirements.txt
alembic==1.13.0
anndata==0.10.3
annoy==1.17.3
anyio==3.7.1
array-api-compat==1.4
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.1
catboost==1.2.2
click==8.1.7
colorlog==6.8.0
contourpy==1.2.0
cycler==0.12.1
databases==0.6.2
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.105.0
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
graphviz==0.20.1
greenlet==3.0.2
h11==0.14.0
h5py==3.10.0
idna==3.6
imap==1.0.0
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
optuna==3.5.0
ormar==0.12.2
packaging==23.2
pandas==2.1.4
patsy==0.5.4
Pillow==10.1.0
plotly==5.18.0
psycopg2==2.9.9
pydantic==1.10.8
PyJWT==2.8.0
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
scanpy==1.9.6
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.12.2
session-info==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.27.0
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
umap-learn==0.5.5
uvicorn==0.24.0.post1